added authorization for all routes

+++ pages/api/daily/delete.ts
@@ -1,3 +1,4 @@
+import authorize from "@/lib/authorize"
 import connectDb from "@/lib/connectDb"
 import dailyModel from "@/models/daily.model"
 import { NextApiRequest, NextApiResponse } from "next"
@@ -12,16 +13,17 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
 			return
 		}
 
-		const { id } = req.body
-		
+		const user = authorize(req, res)
+		if (user) {
+			const { id } = req.body
+			const task = await dailyModel.deleteOne({ _id: id })
 
-		const task = await dailyModel.deleteOne({ _id: id })
-
-		res.status(200).send({
-			success: true,
-			message: "task deleted successfully",
-			task
-		})
+			res.status(200).send({
+				success: true,
+				message: "task deleted successfully",
+				task
+			})
+		}
 
 	} catch (error) {
 		console.log(error)
